# target/jaxrs-analyzer/swagger.json
{
    "swagger":"2.0",
    "info":{
        "version":"0.0.1-SNAPSHOT",
        "title":"jax-rs_Analyzer"
    },
    "host":"exemple.com",
    "basePath":"/",
    "schemes":[
        "http"
    ],
    "paths":{
        "/Coffee":{
            "get":{
                "consumes":[
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"type",
                        "in":"query",
                        "required":true
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        },
                        "schema":{
                            "type":"array",
                            "items":{
                                "$ref":"#/definitions/Coffee"
                            }
                        }
                    }
                }
            }
        },
        "/Coffee/{id}":{
            "get":{
                "consumes":[
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"id",
                        "in":"query",
                        "required":true
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        },
                        "schema":{
                            "$ref":"#/definitions/Coffee"
                        }
                    }
                }
            }
        }
    },
    "definitions":{
        "Coffee":{
            "properties":{
                "strength":{
                    "type":"integer"
                },
                "type":{
                    "type":"string"
                }
            }
        }
    }
}